refactor: optimize project structure
