--- src/components/Error.tsx
@@ -1,11 +1,0 @@
-//import { useState } from 'react';
-
-function Error() {
-  return (
-    <>
-      <p>Error Description</p>
-    </>
-  );
-}
-
-export default Error;
