+++ src/components/ResultsNotFound.tsx
@@ -1,0 +1,5 @@
+function ResultsNotFound() {
+  return <div className="resultsNotFound">Nothing found</div>;
+}
+
+export default ResultsNotFound;
